requirements.txt
langchain
openai
#langchain-core
#langchain-community
langchain-openai
langchain-google-genai
langchain_experimental
openai
anthropic
cohere
streamlit
python-dotenv
beautifulsoup4
faiss-cpu
scrapetube
yt_dlp
pydub
ffmpeg
ffprobe
